+++ sample-outputs/len.c
@@ -1,0 +1,5 @@
+static Py_ssize_t
+list_length(PyListObject *self)
+{
+    return Py_SIZE(self);
+}