+++ sample-outputs/len.py
@@ -1,0 +1,2 @@
+def my_len(x):
+    return len(x)